Add a temporary utility module for BYTEA encoding/decoding
It is temporary in the sense that we're going to move the implementation elsewhere
when we add native support for the BYTEA type to Electric.
